huazie_flea-frame-cache 修复set和put获取memcached主机不一致_MOD

+++ flea-frame-cache/src/main/java/com/huazie/frame/cache/AbstractFleaCache.java
@@ -5,6 +5,7 @@ import com.huazie.frame.common.util.ObjectUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.Date;
 import java.util.HashSet;
 import java.util.Set;
 
@@ -21,11 +22,11 @@ public abstract class AbstractFleaCache implements IFleaCache {
 
     private Set<String> keySet = new HashSet<String>();
     private final String name;  // 缓存主要关键字（用于区分）
-    private final int expire;   // 失效时间（毫秒单位）
+    private final long expire;  // 失效时间(单位：秒)
 
     protected CacheEnum cache;  // 缓存类型
 
-    public AbstractFleaCache(String name, int expire) {
+    public AbstractFleaCache(String name, long expire) {
         this.name = name;
         this.expire = expire;
     }
@@ -81,10 +82,31 @@ public abstract class AbstractFleaCache implements IFleaCache {
         }
     }
 
+    /**
+     * <p> 获取缓存值 </p>
+     *
+     * @param key 缓存键
+     * @return 缓存值
+     * @since 1.0.0
+     */
     protected abstract Object getNativeValue(String key);
 
-    protected abstract void putNativeValue(String key, Object value, int expire);
+    /**
+     * <p> 添加缓存数据 </p>
+     *
+     * @param key    缓存键
+     * @param value  缓存值
+     * @param expire 失效时间（单位：秒）
+     * @since 1.0.0
+     */
+    protected abstract void putNativeValue(String key, Object value, long expire);
 
+    /**
+     * <p> 删除指定缓存数据 </p>
+     *
+     * @param key 缓存键
+     * @since 1.0.0
+     */
     protected abstract void deleteNativeValue(String key);
 
     protected String getNativeKey(String key) {
@@ -95,7 +117,7 @@ public abstract class AbstractFleaCache implements IFleaCache {
         return name;
     }
 
-    public int getExpire() {
+    public long getExpire() {
         return expire;
     }
