+++ flea-frame-cache/src/main/java/com/huazie/frame/cache/AbstractSpringCacheManager.java
@@ -20,7 +20,7 @@ public abstract class AbstractSpringCacheManager extends AbstractTransactionSupp
 
     private static ConcurrentMap<String, AbstractSpringCache> cacheMap = new ConcurrentHashMap<String, AbstractSpringCache>();
 
-    protected Map<String, Integer> configMap = new HashMap<String, Integer>();   // 各缓存的时间Map
+    protected Map<String, Long> configMap = new HashMap<String, Long>();   // 各缓存的时间Map
 
     @Override
     protected Collection<? extends AbstractSpringCache> loadCaches() {
@@ -32,7 +32,7 @@ public abstract class AbstractSpringCacheManager extends AbstractTransactionSupp
     public AbstractSpringCache getCache(String name) {
         synchronized (cacheMap) {
             if (!cacheMap.containsKey(name)) {
-                Integer expire = configMap.get(name);
+                Long expire = configMap.get(name);
                 if (expire == null) {
                     expire = CommonConstants.NumeralConstants.ZERO; // 表示永久
                     configMap.put(name, expire);
@@ -51,9 +51,9 @@ public abstract class AbstractSpringCacheManager extends AbstractTransactionSupp
      * @return 新建的缓存对象
      * @since 1.0.0
      */
-    protected abstract AbstractSpringCache newCache(String name, int expire);
+    protected abstract AbstractSpringCache newCache(String name, long expire);
 
-    public void setConfigMap(Map<String, Integer> configMap) {
+    public void setConfigMap(Map<String, Long> configMap) {
         this.configMap = configMap;
     }
