+++ flea-frame-cache/src/main/java/com/huazie/frame/cache/redis/RedisFleaCacheManager.java
@@ -13,7 +13,7 @@ import com.huazie.frame.cache.AbstractSpringCacheManager;
 public class RedisFleaCacheManager extends AbstractSpringCacheManager {
 
     @Override
-    protected AbstractSpringCache newCache(String name, int expire) {
+    protected AbstractSpringCache newCache(String name, long expire) {
         return null;
     }
